feat: update stdio cmd
